requirements.txt
fastapi
uvicorn
python-dotenv
mlflow==2.16.2
cloudpickle==3.0.0
numpy==1.26.4
pandas==2.2.1
psutil==5.9.0
scikit-learn==1.5.2
scipy==1.13.1
starlette
httpx
loguru==0.7.2